refactor: reorg ContextGui.lua to Player/
